+++ src/amg/azext_amg/commands.py
@@ -13,8 +13,8 @@ def load_command_table(self, _):
         self.command_table['grafana create'] = GrafanaCreate(loader=self)
         self.command_table['grafana delete'] = GrafanaDelete(loader=self)
         self.command_table['grafana update'] = GrafanaUpdate(loader=self)
-        g.custom_command('backup', 'backup_grafana', is_preview=True)
-        g.custom_command('restore', 'restore_grafana', is_preview=True)
+        g.custom_command('backup', 'backup_grafana')
+        g.custom_command('restore', 'restore_grafana')
         g.custom_command('migrate', 'migrate_grafana', is_preview=True)
 
     with self.command_group('grafana dashboard') as g:
@@ -34,14 +34,6 @@ def load_command_table(self, _):
         g.custom_command('query', 'query_data_source')
         g.custom_command('update', 'update_data_source')
 
-    with self.command_group('grafana notification-channel', deprecate_info=self.deprecate()) as g:
-        g.custom_command('list', 'list_notification_channels')
-        g.custom_show_command('show', 'show_notification_channel')
-        g.custom_command('create', 'create_notification_channel')
-        g.custom_command('update', 'update_notification_channel')
-        g.custom_command('delete', 'delete_notification_channel')
-        g.custom_command('test', 'test_notification_channel')
-
     with self.command_group('grafana folder') as g:
         g.custom_command('create', 'create_folder')
         g.custom_command('list', 'list_folders')
@@ -53,11 +45,6 @@ def load_command_table(self, _):
         g.custom_command('list', 'list_users')
         g.custom_show_command('show', 'show_user')
         g.custom_command('actual-user', 'get_actual_user')
-
-    with self.command_group('grafana api-key', deprecate_info=self.deprecate()) as g:
-        g.custom_command('create', 'create_api_key')
-        g.custom_command('list', 'list_api_keys')
-        g.custom_command('delete', 'delete_api_key')
 
     with self.command_group('grafana service-account') as g:
         g.custom_command('create', 'create_service_account')
@@ -72,6 +59,6 @@ def load_command_table(self, _):
         g.custom_command('delete', 'delete_service_account_token')
 
     with self.command_group('grafana integrations monitor') as g:
-        g.custom_command('add', 'link_monitor', is_preview=True)
-        g.custom_command('list', 'list_monitors', is_preview=True)
-        g.custom_command('delete', 'unlink_monitor', is_preview=True)
+        g.custom_command('add', 'link_monitor')
+        g.custom_command('list', 'list_monitors')
+        g.custom_command('delete', 'unlink_monitor')